move fns to src/data/xml_extraction.py for testing and pep8 tidy
